Lv5: 내가 정의한 문제와 해결 과정 (3)
1. CommentAdminService 리팩토링
2. CommentId 존재하지 않을 경우 예외 처리

+++ src/main/java/org/example/expert/domain/comment/service/CommentAdminService.java
@@ -2,6 +2,7 @@ package org.example.expert.domain.comment.service;
 
 import lombok.RequiredArgsConstructor;
 import org.example.expert.domain.comment.repository.CommentRepository;
+import org.example.expert.domain.common.exception.InvalidRequestException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -11,8 +12,12 @@ public class CommentAdminService {
 
     private final CommentRepository commentRepository;
 
+    // 리펙토링 3) commentId 가 없을 경우 오류 반환
     @Transactional
     public void deleteComment(long commentId) {
+        if (!commentRepository.existsById(commentId)) {
+            throw new InvalidRequestException("Comment not found");
+        }
         commentRepository.deleteById(commentId);
     }
 }
